code/requirements.txt
pygame==2.5.2
mutagen==1.47.0
Pillow==10.0.1
RPi.GPIO==0.7.1
spidev==3.6
